ES-878340 Renamed sample addressed the feedbacks ES-883114 Added sample for replace cell content
